Results-exercise view and improvements
