+++ assets/styles/codetest.scss
@@ -69,18 +69,15 @@
     padding-top: 0;
     padding-bottom: 0;
 }
-
-.list-group-item{
-    background-color: #FCFCFC;
+.list-group-item:last-child{
+    border-bottom: 0px;
 }
-
 div.header-col {
     padding-top: 10px;
    /*padding-bottom: 10px;*/
 }
 .response-row {
     border-top: 1px solid #C4C4C4;
-    background-color: #FEF3E1;
 }
 a.response-collapse-link,
 a.response-collapse-link:active,
@@ -89,11 +86,7 @@ a.response-collapse-link:hover {
     color: #111111;
     opacity: 1;
 }
-.selected-row {
-    background-color: #FFC54A;
-}
 .response-badge {
-    background-color: #FFC54A;
     color: #111111;
     font-weight: 900;
     font-size: 14px;
@@ -447,11 +440,6 @@ html{
 .lastElementForm{
     margin-bottom: 15px;
 }
-
-.response-exercise-row {
-    background-color: #FEF3E1;
-}
-
 .display-code{
     font-size: 0.8em;
     font-weight: bold;
