--- views/answer/getAnswersFromExercise.php.twig
@@ -1,18 +1,0 @@
-{% for response in responses %}
-    {% set user = response.user %}
-    <div class="row response-row">
-        <div class="col-sm-2">
-            <h5>{{ user.getDisplayname() }}</h5>
-            <p>{{ response.formattedResponseDate }}</p>
-        </div>
-        <div class="col-sm-8" style="height:300px;overflow: scroll;">
-            <p class="response-text"><pre>{{ response.answer.getAnswerTxt() }}</pre></p>
-        </div>
-        <div class="col-sm-2">
-            <span aria-hidden="true" class="fas fa-thumbs-{{response.answer.getAnswerSuccess() ? 'up' : 'down'}} text-success"></span>
-            {% if response.answer.getAnswerSuccess() != 1 %}
-                {{ include('answer/answer-diffs.php', { 'answer' : response.answer}) }}
-            {% endif %}
-        </div>
-    </div>
-{% endfor %}
