+++ views/answer/results-exercise.php.twig
@@ -3,33 +3,89 @@
 {{ OUTPUT.pageTitle('pages-title.result.by.exercise'|trans, true, false) }}
 {% endblock %}
 {% block content %}
-        <section id="exerciseResponses">
-            <div class="list-group">
-                {% for exercise in exercises %}
-                    {% set exerciseId = exercise.getExerciseId() %}
-                    {% set numberResponses = exercise.getNumberAnswers() %}
-                    <div class="list-group-item response-list-group-item">
-                        <div class="row">
-                            <div class="col-sm-3 header-col">
-                                <a class="h4 response-collapse-link" data-toggle="collapse" data-target="#responses{{ exerciseId }}" onclick="getAnswersFromExercise('{{ exerciseId }}')">
+<div class="results-exercise-page">
+    <div class="panel panel-info">
+        <div class="panel-heading response-panel-header">
+            <div class="row">
+                <div class="row d-flex justify-center items-center font-bold text-lg text-blue-900 h-14">
+                    <div class="col-sm-3 d-flex justify-center items-center">
+                        <h4 class="results-table-hdr">{%trans%}results-table.exercise-number{%endtrans%}</h4>
+                    </div>
+                    <div class="d-flex justify-center items-center col-sm-8">
+                        <h4 class="results-table-hdr">{%trans%}results-table.exercise-statement{%endtrans%}</h4>
+                    </div>
+                    <div class="col-sm-1 d-flex justify-center items-center"></div>                
+                </div>
+        </div>  
+    </div>        
+    <section id="exerciseResponses">
+        <div class="list-group">
+            {% for exercise in exercises %}
+                {% set exerciseId = exercise.getExerciseId() %}
+                {% set answers = exercise.getAnswers() %}
+                {% set numberResponses = exercise.getNumberAnswers() %}
+                <div class="list-group-item response-list-group-item">
+                    <div class="row">
+                        <div class="row d-flex min-h-12">
+                            <div class="col-sm-3 d-flex justify-center items-center">
+                                <a class="h4 response-collapse-link" data-toggle="collapse" data-target="#responses{{ exerciseId }}">
                                     {%trans%}general.exercise{%endtrans%} {{ exercise.getExerciseNum() }}
                                     <span class="fa fa-chevron-down rotate" aria-hidden="true"></span>
                                 </a>
                             </div>
-                            <div class="col-sm-offset-1 col-sm-8 header-col">
+                            <div class="col-sm-8 d-flex justify-center items-center">
                                 <div class="flx-cntnr flx-row flx-nowrap flx-start">
                                     {% autoescape false %}
                                     <span class="flx-grow-all">{{ exercise.getTitle() }}</span>
                                     {% endautoescape %}
-                                    <span class="badge response-badge">{{ numberResponses }}</span>
                                 </div>
                             </div>
-                            <div id="responses{{ exerciseId }}" class="col-xs-12 results-collapse collapse">
-                            </div>
-                        </div>
+                            <div class="col-sm-1 d-flex justify-center items-center"></div>         
+                        </div>    
+                        <div id="responses{{ exerciseId }}" class="col-xs-12 results-collapse collapse" style="background-color: #d6ecff73;">
+                            {% for response in responses %}
+                                {% set user = response.user %}
+                                {% set answer = user.getAnswerForExercise(exercise.getExerciseId(), exercise.getCtId()) %}
+                                {% if answer.answerId == response.answer.answerId %}
+                                <div class="row response-row d-flex" style="min-height:70px;">
+                                    <div class="col-sm-3 d-flex self-center justify-center">
+                                        <div>
+                                            <h5>{{ user.getDisplayname() }}</h5>
+                                            <p>{{ response.formattedResponseDate }}</p>
+                                        </div>
+                                    </div>
+                                    <div class="col-sm-8 d-flex justify-center items-center">
+                                        <div class="col-xs-4 d-flex self-center justify-center">
+                                            <button id="btnModalCode{{response.answer.getAnswerId()}}" onclick="showModal({{response.answer.answerId}})" class="display-code mr-3">
+                                                {%trans%}results-table.show-code{%endtrans%}
+                                            </button>
+                                        </div>              
+                                        <div id="modalCode{{response.answer.getAnswerId()}}" class="code-modal">
+                                            <div class="code-modal-content row">
+                                                <div class="font-bold text-lg mt-2">
+                                                    <div class="col-xs-6 d-flex">{%trans%}general.exercise{%endtrans%} {{ exercise.getExerciseNum() }}</div>
+                                                    <div class="col-xs-6 d-flex justify-end">{{ user.getDisplayname() }}</div>
+                                                </div>
+                                                <div class="col-xs-12 h-4 border-b border-blue-900"></div>
+                                                <div class="col-xs-12 d-flex mt-5 whitespace-pre text-blue-900 bg-gray-100 rounded-md p-5"><a>{{ response.answer.getAnswerTxt() }}</a></div> 
+                                            </div>
+                                        </div>
+                                    </div>
+                                    <div class="col-sm-1 d-flex justify-center items-center">
+                                        <span aria-hidden="true" class="{{answer.getAnswerSuccess() ? 'fa fa-check-circle-o' : 'fa fa-times-circle-o'}} d-flex self-center justify-center"></span> 
+                                        {% if response.answer.getAnswerSuccess() != 1 %}
+                                            {{ include('answer/answer-diffs.php', { 'answer' : response.answer}) }}
+                                        {% endif %}
+                                    </div>
+                                </div>
+                                {% endif %}
+                            {% endfor %} 
+                        </div> 
                     </div>
-                {% endfor %}
-            </div>
-        </section>
-    </div>
+                </div>
+            {% endfor %}
+        </div>
+    </section>
+</div>
+        
 {% endblock %}
